+++ lwrb/src/include/lwrb/lwrb.h
@@ -29,11 +29,12 @@
  * This file is part of LwRB - Lightweight ring buffer library.
  *
  * Author:          Tilen MAJERLE <tilen@majerle.eu>
- * Version:         v3.0.0-rc1
+ * Version:         v3.1.0
  */
 #ifndef LWRB_HDR_H
 #define LWRB_HDR_H
 
+#include <stddef.h>
 #include <stdint.h>
 #include <string.h>
 
@@ -47,11 +48,23 @@ extern "C" {
  * \{
  */
 
-#ifdef LWRB_DISABLE_ATOMIC
-typedef unsigned long lwrb_ulong_t;
-#else
+#if !defined(LWRB_DISABLE_ATOMIC) || __DOXYGEN__
 #include <stdatomic.h>
-typedef atomic_ulong lwrb_ulong_t;
+
+/**
+ * \brief           Atomic type for size variable.
+ * Default value is set to be `unsigned 32-bits` type
+ */
+typedef atomic_ulong lwrb_sz_atomic_t;
+
+/**
+ * \brief           Size variable for all library operations.
+ * Default value is set to be `unsigned 32-bits` type
+ */
+typedef unsigned long lwrb_sz_t;
+#else
+typedef unsigned long lwrb_sz_atomic_t;
+typedef unsigned long lwrb_sz_t;
 #endif
 
 /**
@@ -74,48 +87,56 @@ struct lwrb;
  * \param[in]       evt: Event type
  * \param[in]       bp: Number of bytes written or read (when used), depends on event type
  */
-typedef void (*lwrb_evt_fn)(struct lwrb* buff, lwrb_evt_type_t evt, size_t bp);
+typedef void (*lwrb_evt_fn)(struct lwrb* buff, lwrb_evt_type_t evt, lwrb_sz_t bp);
+
+/* List of flags */
+#define LWRB_FLAG_READ_ALL  ((uint16_t)0x0001)
+#define LWRB_FLAG_WRITE_ALL ((uint16_t)0x0001)
 
 /**
  * \brief           Buffer structure
  */
 typedef struct lwrb {
     uint8_t* buff;  /*!< Pointer to buffer data. Buffer is considered initialized when `buff != NULL` and `size > 0` */
-    size_t size;    /*!< Size of buffer data. Size of actual buffer is `1` byte less than value holds */
-    lwrb_ulong_t r; /*!< Next read pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
-    lwrb_ulong_t w; /*!< Next write pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
+    lwrb_sz_t size; /*!< Size of buffer data. Size of actual buffer is `1` byte less than value holds */
+    lwrb_sz_atomic_t r; /*!< Next read pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
+    lwrb_sz_atomic_t w; /*!< Next write pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
     lwrb_evt_fn evt_fn; /*!< Pointer to event callback function */
 } lwrb_t;
 
-uint8_t lwrb_init(lwrb_t* buff, void* buffdata, size_t size);
+uint8_t lwrb_init(lwrb_t* buff, void* buffdata, lwrb_sz_t size);
 uint8_t lwrb_is_ready(lwrb_t* buff);
 void lwrb_free(lwrb_t* buff);
 void lwrb_reset(lwrb_t* buff);
 void lwrb_set_evt_fn(lwrb_t* buff, lwrb_evt_fn fn);
 
 /* Read/Write functions */
-size_t lwrb_write(lwrb_t* buff, const void* data, size_t btw);
-size_t lwrb_read(lwrb_t* buff, void* data, size_t btr);
-size_t lwrb_peek(const lwrb_t* buff, size_t skip_count, void* data, size_t btp);
+lwrb_sz_t lwrb_write(lwrb_t* buff, const void* data, lwrb_sz_t btw);
+lwrb_sz_t lwrb_read(lwrb_t* buff, void* data, lwrb_sz_t btr);
+lwrb_sz_t lwrb_peek(const lwrb_t* buff, lwrb_sz_t skip_count, void* data, lwrb_sz_t btp);
+
+/* Extended read/write functions */
+uint8_t lwrb_write_ex(lwrb_t* buff, const void* data, lwrb_sz_t btw, lwrb_sz_t* bw, uint16_t flags);
+uint8_t lwrb_read_ex(lwrb_t* buff, void* data, lwrb_sz_t btr, lwrb_sz_t* br, uint16_t flags);
 
 /* Buffer size information */
-size_t lwrb_get_free(const lwrb_t* buff);
-size_t lwrb_get_full(const lwrb_t* buff);
+lwrb_sz_t lwrb_get_free(const lwrb_t* buff);
+lwrb_sz_t lwrb_get_full(const lwrb_t* buff);
 
 /* Read data block management */
 void* lwrb_get_linear_block_read_address(const lwrb_t* buff);
-size_t lwrb_get_linear_block_read_length(const lwrb_t* buff);
-size_t lwrb_skip(lwrb_t* buff, size_t len);
+lwrb_sz_t lwrb_get_linear_block_read_length(const lwrb_t* buff);
+lwrb_sz_t lwrb_skip(lwrb_t* buff, lwrb_sz_t len);
 
 /* Write data block management */
 void* lwrb_get_linear_block_write_address(const lwrb_t* buff);
-size_t lwrb_get_linear_block_write_length(const lwrb_t* buff);
-size_t lwrb_advance(lwrb_t* buff, size_t len);
+lwrb_sz_t lwrb_get_linear_block_write_length(const lwrb_t* buff);
+lwrb_sz_t lwrb_advance(lwrb_t* buff, lwrb_sz_t len);
 
 /* Search in buffer */
-uint8_t lwrb_find(const lwrb_t* buff, const void* bts, size_t len, size_t start_offset, size_t* found_idx);
-size_t lwrb_overwrite(lwrb_t* buff, const void* data, size_t btw);
-size_t lwrb_move(lwrb_t* dest, lwrb_t* src);
+uint8_t lwrb_find(const lwrb_t* buff, const void* bts, lwrb_sz_t len, lwrb_sz_t start_offset, lwrb_sz_t* found_idx);
+lwrb_sz_t lwrb_overwrite(lwrb_t* buff, const void* data, lwrb_sz_t btw);
+lwrb_sz_t lwrb_move(lwrb_t* dest, lwrb_t* src);
 
 /**
  * \}